Function calling with scoped arguments
and some refactoring w.r.t. the Result / Error stuff.
Error now is an own enum and not an Object so we do not have
Result<Object, Object> anymore.

Also got rid of the evaluator crate since that is all handled in
the CoerceObject trait. Evaluator was only passing through.